EE-400: Move genesis related logic to genesis.rs.
