+++ execution-engine/comm/src/engine_server/mod.rs
@@ -7,9 +7,10 @@ use std::time::Instant;
 use common::key::Key;
 use common::value::account::{BlockTime, PublicKey};
 use common::value::U512;
-use execution_engine::engine_state::{EngineState, GenesisResult};
+use execution_engine::engine_state::EngineState;
 use execution_engine::engine_state::error::Error as EngineError;
 use execution_engine::engine_state::execution_result::ExecutionResult;
+use execution_engine::engine_state::genesis::GenesisResult;
 use execution_engine::execution::{Executor, WasmiExecutor};
 use execution_engine::tracking_copy::QueryResult;
 use shared::logging;
